requirements.txt
# pip install -r requirements.txt

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
pandas>=1.1.4
seaborn>=0.11.0
SimpleITK==2.1.1.2
pydicom==2.3.0
timm==0.5.4
scikit-learn==0.24.2
gdown==4.4.0
scikit-image==0.19.2